+++ opportunities-service/src/main/java/com/ironhack/opportunitiesservice/controller/interfaces/IOpportunityController.java
@@ -14,15 +14,15 @@ import java.util.Optional;
 public interface IOpportunityController {
 
     //Get method: get a list with all the opportunities
-    List<OpportunityDTO> getOpportunitiesBy(Optional<Integer> salesRepId, Optional<Status> status,
+    List<OpportunityDTO> getOpportunitiesBy(Optional<Long> salesRepId, Optional<Status> status,
                                             Optional<String> country, Optional<String> city,
                                             Optional<Industry> industry);
     //Get method: get a opportunityDTO by id
-    OpportunityDTO getOpportunityDTOById(int id);
+    OpportunityDTO getOpportunityDTOById(Long id);
     //Post method: create a new opportunity
     Opportunity createOpportunity(OpportunityDTO opportunityDTO);
     //Patch method: update the status of a opportunity
-    void updateOpportunityStatus(int id, OpportunityStatusDTO opportunityStatusDTO);
+    void updateOpportunityStatus(Long id, OpportunityStatusDTO opportunityStatusDTO);
     //Patch method: update the accountId of a opportunity
-    void updateOpportunityAccountId(int id, AccountIdDTO accountIdDTO);
+    void updateOpportunityAccountId(Long id, AccountIdDTO accountIdDTO);
 }
