+++ opportunities-service/src/main/java/com/ironhack/opportunitiesservice/service/interfaces/IOpportunityService.java
@@ -14,11 +14,11 @@ public interface IOpportunityService {
     //Get method: get all the opportunities
     List<OpportunityDTO> getAllOpportunities();
     //Get method: get an Opportunity by id
-    OpportunityDTO getOpportunityById(int id);
+    OpportunityDTO getOpportunityById(Long id);
     //Get method: get an opportunity list by sales rep
-    List<OpportunityDTO> getOpportunitiesBySalesRep(int salesRepId);
+    List<OpportunityDTO> getOpportunitiesBySalesRep(Long salesRepId);
     //Get method: get an opportunity list by sales rep and status
-    List<OpportunityDTO> getOpportunitiesBySalesRepAndStatus(int salesRepId, Status status);
+    List<OpportunityDTO> getOpportunitiesBySalesRepAndStatus(Long salesRepId, Status status);
     //Get method:get an opportunity list by country
     List<OpportunityDTO> getOpportunitiesByCountry(String country);
     //Get method:get an opportunity list by country and status
@@ -34,9 +34,9 @@ public interface IOpportunityService {
     //Post method: create a new Opportunity
     Opportunity createOpportunity(OpportunityDTO opportunityDTO);
     //Patch method: modify the status of the opportunity
-    void updateOpportunityStatus(int id, OpportunityStatusDTO opportunityStatusDTO);
+    void updateOpportunityStatus(Long id, OpportunityStatusDTO opportunityStatusDTO);
     //Patch method: modify the accountId of the opportunity
-    void updateOpportunityAccountId(int id, AccountIdDTO accountIdDTO);
+    void updateOpportunityAccountId(Long id, AccountIdDTO accountIdDTO);
 
 
 }
